backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
redis==5.2.0
aiofiles==24.1.0
pillow==11.0.0
python-dotenv==1.0.1
httpx==0.27.2
emails==0.6.0
psycopg2-binary==2.9.10
email-validator==2.2.0
passlib==1.7.4
bcrypt==4.0.1
